+++ exporter/signalfxexporter/dimensions/metadata.go
@@ -22,7 +22,7 @@ import (
 	"go.opentelemetry.io/collector/component/componenterror"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/signalfxexporter/translation"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/metadata"
+	metadata "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/experimentalmetricmetadata"
 )
 
 // MetadataUpdateClient is an interface for pushing metadata updates
