+++ exporter/signalfxexporter/dimensions/requests.go
@@ -1,4 +1,4 @@
-// Copyright -c Google LLC
+// Copyright OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
