+++ pkg/experimentalmetricmetadata/README.md
@@ -1,3 +1,3 @@
-# Metric Metadata
+# Experimental Metric Metadata
 
 This module is experimental and its functionality and interfaces are not guaranteed to be stable or permanent.
