+++ pkg/experimentalmetricmetadata/go.mod
@@ -1,0 +1,5 @@
+module github.com/open-telemetry/opentelemetry-collector-contrib/pkg/experimentalmetricmetadata
+
+go 1.14
+
+require github.com/stretchr/testify v1.7.0
