+++ pkg/experimentalmetricmetadata/metadata_test.go
@@ -1,4 +1,4 @@
-// Copyright -c Google LLC
+// Copyright 2021 OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package metadata
+package experimentalmetricmetadata
 
 import (
 	"testing"
