--- pkg/metadata/go.mod
@@ -1,5 +1,0 @@
-module github.com/open-telemetry/opentelemetry-collector-contrib/pkg/metadata
-
-go 1.14
-
-require github.com/stretchr/testify v1.7.0
