+++ receiver/k8sclusterreceiver/collection/nodes.go
@@ -25,7 +25,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/common/testing/util"
-	metadataPkg "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/metadata"
+	metadataPkg "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/experimentalmetricmetadata"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/k8sclusterreceiver/utils"
 )
